third_party/blink/web_tests/media/W3C/audio/error/error_onerror_called_on_bogus_source-expected.txt

audioElement.onerror is triggered on bogus source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "4" is "4"
PASS successfullyParsed is true

TEST COMPLETE


spec reference
